move ConnectionTest to extension
