+++ duckbot/util/connection_test/__init__.py
@@ -1,0 +1,4 @@
+from .connection_test import ConnectionTest
+
+def setup(bot):
+    bot.add_cog(ConnectionTest(bot))
