+++ tests/util/connection_test/test_connection_test.py
@@ -1,0 +1,10 @@
+import pytest
+import mock
+from duckbot.util.connection_test import ConnectionTest
+
+
+@pytest.mark.asyncio
+async def test_connection_success_shuts_down_bot(bot):
+    clazz = ConnectionTest(bot)
+    await clazz.connection_success()
+    assert bot.is_closed()
